Add a diffusion pass
It turns out that the print version of GPU Gems goes into a lot more
detail than the online one. In particular, it discusses the diffusion
pass, which I’ve seen mentioned in the screensaver, but is usually
absent from most fluid sims I’ve encountered online. This pass is
exactly what I need. It dissipates the velocity field, making it much
more well-behaved. And it reuses the Jacobi solver (with some minor
modification).

A few more changes:

- Modify the Jacobi solver to work on `vec2`. This let’s us reuse the
  shader for the diffusion pass.

- Pass `texel_size` to the fluid vertex shader. This determines the size
  of our sampling area. This will later have to be modified to account
  for aspect ratio.

 - Set the `alpha`/`rBeta` parameters in the pressure pass according to
   the suggested values in GPU Gems.

 - Remove viscosity from the divergence pass. I don’t think we need it
   here right now.