Draw endpoints on the lines (line caps)
This is a bit tricky. We draw a full circle when the lines are small to
highlight the ends. But when the lines are large, we draw only half.
The problem is with the additive alpha blend. I’m not yet sure how to
solve this, but it might require rendering the lines + endpoints to a
render buffer, and then compositing them at the end.