Improve performance by 300%
Yeah, so `get_parameter` is kinda slow. Here is was, thinking that the
pressure pass was just super expensive, when I realised that most of
the rendering time was spent on this. Profiling FTW!

I thought I was being smart by saving/restoring the viewport, but this
turned out to be a huge bottleneck. It makes sense now though. I read a
while back that `get_parameter` causes everything to stall while the
CPU/GPU synchronise. Don’t quote me on the details here. The point is
that it’s slow.