Improve texture zero-ing performance
I’m honestly not sure if this actually works.

In the previous iteration, we were creating and buffering an array of
zeros before writing to the texture. This is slow and inefficient. We
could cache the buffer of zeros to gain some time here.

I’ve seen people use a fragment shader to do rewrite textures, which
makes sense. But should’nt `clear_color` accomplish the same thing?