Fix fluid grid size
I thought I understood what `grid_size` was for, but it appears not.
Looks like `1.0` is the right value for us here after all, as `1.0 /
grid_width` just kills the fluid and turns it into a homogeneous blob.
And while it makes sense that the smaller value reduces advection, I
don’t quite grok why it looks fine when advecting dye in a color
texture.

GPU Gems talks about this parameter scaling the velocity into grid space
because the texture coordinates are in the wrong range. But then in the
code, it just sets the parameter to 1.