Reduce the processing time in Jacobi iterations by 50%
Setting the uniforms in a loop is very expensive. In particular, a lot
of the time is taken up by HashMap and String. Most of these uniforms
are either static or need to be set just once per tick.

As part of this, uniforms are now passed as a reference to the draw
functions. Vertex buffers should probably follow suite.

Quite a few methods in the rendering engine that work with uniforms no
longer need lifetimes. It turns out we were taking ownership of
textures when setting them, so we ended up having to pepper all of our
rendering functions with lifetimes. The solution was to use `ref` when
pattern matching.