Implement 6-segment color wheel
You can set up to 6 colors, each getting π/3 or 60° of color wheel. The
color of the line is determined by the angle of the line, plus extra
based on its length. This extra bit gives us more color variation and
natural-looking highlights in areas where the velocity field is strong.

I think the natural next step here is to allow the color wheel to be
asymmetrical. Imagine we’re extracting colors from images; the dominant
colors aren’t going to be equally represented. So it would make sense
to give each color an amount of space on the color wheel proportional
to its incidence in the image.