Hide scrollbars and request dark mode
There’s this new `color-scheme` property that lets you request
dark mode. This is useful for browsers on windows where the scrollbars
are not very pleasing to look at.
And styling scrollbars is still a pain. Chrome starts lagging with
styled scrollbars, and there’s no easy way to target a specific
platform.