Fix projection scaling bug when sampling fluid velocity
I had at some point decided to pre-multiply the projection and view
(scale) matrices, and I ended up using this matrix to project the
basepoints to clip space when sampling from the velocity texture. The
issue here is that we were now sampling the entire velocity texture
using just the points in the viewport, instead of using all of the
basepoints. That’s why, despite zooming in, you could see the edges of
the fluid simulation.

The solution is: split the view and projection matrices; use the
projection matrix when sampling the velocity fluid; and use both
matrices when drawing the lines and endpoints. Sample first, scale
later.