Fix velocity blending
So far I’ve been treating the noise texture like acceleration, but I
don’t think I’m doing it quite right. I think I need to add an easing
curve to make it work, because it’s a bit too jittery otherwise,
particularly when increasing the advection speed. For now, it’s way
easier to treat the noise as velocity and just blend it in.