Build entire project with nix
The biggest challenge here was actually getting a node_modules folder.
I ended up using mkYarnPackage to make this work for a couple of
reasons:

- pnpm2nix would‘ve been perfect, but it‘s outdated. I almost got it
working, but it fails because of a hardcoded lock file version limit.
Seems to work fine if you manually lower the declared version in the
lock file. Obviously, that‘s not a solution.

- node2nix works, but requires building a separate derivation to get
access to the generated nix expressions. That‘s really annoying! We‘d
also need a package-lock.json if we want locked dependencies. In
general, the API is a bit weird. It‘s really intended to be used as a
standalone CLI.

That leaves us with yarn2nix, which has actually been merged into nix.
No issues with it, other than having to add a yarn.lock. Ah, back to
yarn we go.

Also, shell hooks run from flakes get the flake‘s directory and not the
directory of the nix expression they‘re in. Weird!