Fix nix build in CI
Elm tries to download a package list during the build, but derivations
aren‘t allowed to do that in order to keep things reproducible. We can
use elm2nix to download this registry ahead of time and compute the
hashes for the packages. Nix can then ensure that we’re not downloading
anything unexpected.