Animate line width with changes to velocity
In the original screensaver, you’ll often see long lines thin out when
receeding, and the opposite: shorts ones rapidly thickening before
extending. I think what’s happening is that the width is being somehow
modified by the velocity, but the details are unclear.

I‘ve managed to replicate the effect somewhat by tracking the width
over time and using the velocity to vary it. It’s almost like the width
is also advected by the fluid, instead of being coupled to the length
of the line.