Use uniform buffers for drawer
You can’t bind a buffer to the vertex array object and then write to
it in a transform feedback pass. I feel like it’s better to write the
feedback to a dedicated buffer and the copy the contents. This
sidesteps the pain of having to constantly change the vertex bindings
on every frame.