Highlight endpoints of shorter lines with a bright circle
This is essentially the last must-have feature from the original. This
has been gnawing at my conscience for a while now. Frankly, I just
wasn’t sure how to get it done!

The main issue is: how do you fade in the lower circle that overlaps the
line with additive blending enabled? The lower circle is always going
to be brighter than the upper one because of that overlap.

The solution is surprisingly simple. Take the color of the line,
premultiply by the opacity, and subtract from the base color of the
endpoint. Done! What you end up with is an inverse relationship between
the brightness of the line and the endpoint. As the line gets brighter,
the endpoint dims, and vice-versa.

There is one downside though. You can’t reduce the opacity of the line
too much, otherwise the lower endpoint starts to stand out. I feel like
there is a solution to this, but I don’t see it right now.

Also, remember that the selective coloring in the fragment shader means
that there’s aliasing at the boundary of the two endpoint halves. It
becomes quite visible at large line widths. The original also exhibits
this artifacting, which is why I’m pretty confident it also does this
selective coloring. I do recall the artifacting was gone when rendering
an actual semicircle. So, perhaps, drawing two passes of semicircles
will reduce the aliasing? Or at least be smoothed out by the MSAA pass?