Add minimum noise threshold when blending noise
I’ve been struggling recently with the look of the simulation. I’d get
the noise profile just right only to notice that straight flows of
lines (not in curl noise hotspots) were no longer straight. They’d be
either converging or diverging depending on the direction of advection.

This was especially visible with forward advection: the curl was just
like the original, but the lines constantly looked like they were
diverging.

What I realise now is that I’ve been applying the noise uniformly to
the velocity field. What I actually want are hotspots where the velocity
field curls and the rest of the field does its own thing. Basically,
let the simulation do its thing! This threshold is like a high-pass
filter.