Use logical screen size to compute the line grid
Just because a screen has twice the number of pixels does not mean that
we need to draw twice the number of lines. A 1280x800x2 should have
fewer lines than a 1920x1080x1 because the first screen is probably
smaller, even though they share a similar number of physical pixels.