Improve scaling, especially on smaller devices
This is the third rewrite of this logic! The main issue was that the
grid size was way too small on mobile, so the fluid looked chaotic. We
were also working in physical screen space, which really complicated a
lot of our calculations.

The solution is to work in logical space and treat it like a universal
point space. This is similar to how I used to do it, but back then the
canvas was a fixed size; now we ask the context for the size. We set a
minimum size for this space to avoid weird fluid behaviour
(hello mobile), and then scale it to the physical screen size, while
respecting the physical screen’s aspect ratio.

We heed to Apple’s advice and use physical sizes directly instead of
passing around pixel ratios. That way we correctly support scaled
displays on macOS. In the browser version, we just compute the physical
size ourselves.

I should probably write a struct for these things...