Adjust the fluid size to the proportions of the screen
There’s only so much you can do to avoid distortion when stretching a
square fluid texture onto a rectangular screen. We set a target size
and adjust one of the sides to match the screen ratio. We also resize
the fluid, copying over the previous contents to the new texture, when
resizing the screen. Looks great!