Rework fluid advection and how we generate curl
Previously, I was forcing the fluid to curl when injecting velocity.
While that does work, the fluid can end up curling TOO much sometimes.
Turns out we can just add the noise directly (as acceleration) and do
some magic with advection.

The magic is the following: we advect the fluid forward and in reverse.
The forward pass is the usual Jos Stam advection pass and represents
the state of the fluid at t + 1. The reverse pass tells us what’s going
on with the fluid ahead of the current velocity vector. If the fluid
ahead slows down, we slow down the next state by the current velocity.
If, instead, it speeds up or curls, then we give the next state a
boost.

Additionally:

- Reduce pressure and divergence textures to a single channel. Less
  memory. Yay!
- Extract diffusion into a separate program.