Fix shader compilation on Windows
`sampler2D` is not a valid type when setting precision in core330.
It is in the ES version though, which makes things very confusing. The
proper solution here is to stop writing GLSL directly and instead
write either SPIR-V or WGSL, compiling to whichever GLSL version we
need with naga.