Revert to “default” power preference in WebGL
Requesting “high-performance” in Chrome on dual-GPU macs absolutely
destroys performance for some reason. This seems like a regression
in the OpenGL backend. The new Metal backend works just fine.

The April 2022 WebGL notes mention that GPU switching is broken
right now, but it’s unclear which backend they are referring to.
https://www.khronos.org/assets/uploads/developers/presentations/WebGL_WebGPU_Updates_Apr22.pdf