Simplify the update flow when resizing the grid
It turns out you can send new data to the same buffer handle and the
driver will allocate a new data block for it. Unsurprisingly, I’d never
even tried to do this before and had just assumed you’d gave to delete
the buffer and create a new one. This means I don’t need to redo all the
vertex buffers when resizing. Hurray!

I think, in retrospect, I’m not at fault for not knowing this. I’ve
never seen a tutorial or sample app that describes what to do when the
size of your buffer changes; this is a state management issue, not a
graphics problem.