Fix nix builds again
Well, something changed and now the fake elm bin from node_modules takes
precedence over the binary. I have no idea how the order in the PATH is
set for buildInputs. Simply changing the order of the buildInputs did
nothing, so I’m not going to spend time on this.

Webpack now accepts an optional env variable with the path to elm
compiler.