Implement anti-aliasing in the line and endpoint fragment shaders
I’ve been using MSAA for a while now to combat aliasing. While it
obviously works very well, it’s very expensive, even at the lowest
sample rate of 2. We’re very GPU-bound on lower-end hardware, so adding
MSAA on top of that in the web version was a painful decision. On my
modest machine, MSAA doubles the GPU usage to 50%.

TIL that you can do a cheap antialiasing pass in the fragment shader
when drawing simple shapes. As long as you know where the edges are,
you can use smoothstep to blend the edge with a tiny offset
(fwidth). And drawing simple shapes is exactly what we’re doing — the
whole render pass is just rectangles and circles!

Finally, a use for the `fwidth` function!