Normalise grid coordinates and more
I’ve been using screen-space coordinates for a while now. The reasoning
was that it’d be easier to scale things for different screen sizes.
This proved to be sort of true, but at the cost of complexity in the
shaders, where I’d have to constantly juggle 2 different projection
matrices (one for the fluid, and one for the clip space).

This commit switches the basepoint grid to normalized coordinates, with
dynamic scaling for line widths and lengths.

Endpoints are now quads, instead of triangle fans. Our antialising pass
basically draws circles anyways, so we might as well simplify.

I also think I know why the fluid was so fast before. I was using the
screen-space uLineLength in the velocity calculation...