Fix place lines pass in headless context
Guess what the issue was? Yeah, transform feedback. It turns out that it
requires a backing store to work, despite the fact that the rasterizer
is disabled and we’re definitely not drawing any pixels. It also turns
out that you can bind a 1x1 pixel renderbuffer and it’ll just work.
Ridiculous.