fluid: fix advection bias to upper right corner
Turns out the linear sampling in the diffusion step was adding a
positive bias all along. You can actually see it by sliding the
diffusion iterations slider from 0: the fluid expands to the top right.
What it should do is expand uniformly in all directions.
And here I was, pointing fingers at the floor function in the
adjust_advection routing...

I'm not sure what the issue is, but I've fixed it by switching to
nearest sampling. We now have visual parity with the GL version.
Finally!